Fix sourceforge download URL for SBCL
Altough the -L flag is passed to curl to handle redirects, this
sometimes fails for unknown reasons, a a non-bz2 archive ends up being
downloaded, failing later in unpack().

This fix merely attempts to minimize that failure in the case of SBCL.

Another option would be to integrate the URL retry loop in get() with
unpack(), perhaps in a new get_and_unpack() function that checks for
success of both the download and unpacking.